+++ src/Entity/Item.php
@@ -266,10 +266,10 @@ class Item extends Entity implements StatementListProvider {
 	/**
 	 * @since 1.0
 	 *
-	 * @param StatementList $statements
+	 * @param StatementList|null $statements
 	 */
-	public function setStatements( StatementList $statements ) {
-		$this->statements = $statements;
+	public function setStatements( StatementList $statements = null ) {
+		$this->statements = $statements ?: new StatementList();
 	}
 
 	/**